@@ -1,14 +1,12 @@
-namespace WorkoutTracker.Core.Models{
+namespace WorkoutTracker.Core.Entities;
 
-    public class Workout{
-        public string ID {get; set;}
-        public int StartTime {get; set;}
+public class Workout{
+    public int ID {get; set;}
+    public DateTime StartTime {get; set;}
 
-        public required string Name {get; set;}
+    public required string Name {get; set;}
 
-        public string? Note {get;set;}
+    public string? Note {get;set;}
 
-        //public int ProgramID {get; set;}
-    }
-
+    //public int ProgramID {get; set;}
 }
